Removed unnecessary metadata requests. Removed duplicate metadata request on files added to directory.
The first were accidentally introduced in fd7e2bed87d03427906ffeba1f30ff6422e6e242.
The second were due to the fact that we needed to get the metadata for the file type, but then also needed it again once we opened the image.
With this, we save the metadata if we already had it.