Fixed the zoom moving sideways on some images.
It was a symptom of how we treated out of bound uv coordinates.
We were supposed to crop them to `[0, 1] x [0, 1]` and adjust the ui rectangle depending on the cropped parts.
While the translation of the ui worked, the scaling was done weirdly.
With this commit, we now do both scale and translation in a single pass.
We used to need to check against uvs that were `< 0` or `> 1`, but since we now cap the screen at those, except for when the uvs are bigger than 1, we can just check `width,height > 1`.