Added a thread to load fields without blocking in the main thread. Renamed `update_` methods in `DirEntry` with `load_`.
This is mainly useful for getting the size of an entry.
Until now, we blocked the main thread loading it, because we had no other choice, but now we can just get the value and send it to be loaded if it's unloaded.
It's still a bit in the air whether having this tight dependency between `DirReader` and `DirEntry` is a good idea, but at the very least `DirEntry` doesn't depend directly on `DirReader`, just on the fact that there exists a channel to send requests for loading.

The rename was a drive-by that just coincided with this and the previous commit. Realistically it should've been it's own commit.